Feat: v2 user managment

+++ src/main/java/pe/edu/upc/center/workstation/userManagment/domain/model/queries/user/GetUserByIdQuery.java
@@ -1,4 +1,4 @@
 package pe.edu.upc.center.workstation.userManagment.domain.model.queries.user;
 
-public record GetUserByIdQuery(int userId) {
+public record GetUserByIdQuery(Long userId) {
 }
